fix import of 'testing' in production code
this was due to importing ./utils/testutils package in production code in order
to use ProjectPath, which is why it was moved in ./utils
