+++ utils/functions.go
@@ -2,7 +2,10 @@
 package utils
 
 import (
+	"errors"
 	"os"
+	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/ironsmile/nedomi/types"
@@ -22,4 +25,23 @@ func IsMetadataFresh(obj *types.ObjectMetadata) bool {
 	// This package can be made to work with a precision of one second and never
 	// call time.Now more than that.
 	return time.Unix(obj.ExpiresAt, 0).After(time.Now())
+}
+
+// ProjectPath returns a path to the project source as an absolute directory name.
+func ProjectPath() (string, error) {
+	gopath := os.ExpandEnv("$GOPATH")
+	relPath := filepath.FromSlash("src/github.com/ironsmile/nedomi")
+	for _, path := range strings.Split(gopath, ":") {
+		rootPath := filepath.Join(path, relPath)
+		entry, err := os.Stat(rootPath)
+		if err != nil {
+			continue
+		}
+
+		if entry.IsDir() {
+			return rootPath, nil
+		}
+	}
+
+	return "", errors.New("Was not able to find the project path")
 }
