+++ utils/testutils/test_helpers.go
@@ -1,11 +1,8 @@
 package testutils
 
 import (
-	"errors"
 	"io/ioutil"
 	"os"
-	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 )
@@ -38,23 +35,4 @@ func ShouldntFail(t testing.TB, errors ...error) {
 			t.Fatalf("An unexpected error occured in statement %d: %s", idx+1, err)
 		}
 	}
-}
-
-// ProjectPath returns a path to the project source as an absolute directory name.
-func ProjectPath() (string, error) {
-	gopath := os.ExpandEnv("$GOPATH")
-	relPath := filepath.FromSlash("src/github.com/ironsmile/nedomi")
-	for _, path := range strings.Split(gopath, ":") {
-		rootPath := filepath.Join(path, relPath)
-		entry, err := os.Stat(rootPath)
-		if err != nil {
-			continue
-		}
-
-		if entry.IsDir() {
-			return rootPath, nil
-		}
-	}
-
-	return "", errors.New("Was not able to find the project path")
 }
